+++ cobib/commands/init.py
@@ -4,6 +4,7 @@ import argparse
 import logging
 import os
 import sys
+import warnings
 
 from cobib.config import CONFIG
 from .base_command import ArgumentParser, Command
@@ -26,14 +27,20 @@ class InitCommand(Command):
         LOGGER.debug('Starting Init command.')
         parser = ArgumentParser(prog="init", description="Init subcommand parser.")
         parser.add_argument('-f', '--force', action='store_true',
-                            help="force initializtion of database. Warning: this overwrites any " +
-                            "existing file!")
+                            help="DEPRECATED! This argument will be removed in v2.6")
 
         try:
             largs = parser.parse_args(args)
         except argparse.ArgumentError as exc:
             print("{}: {}".format(exc.argument_name, exc.message), file=sys.stderr)
             return
+
+        if largs.force:
+            msg = 'The "force" argument has been deprecated and will be removed in v2.6\n' + \
+                  'If you would like to overwrite your current database, please edit the file ' + \
+                  'manually.'
+            warnings.warn(msg, DeprecationWarning)
+            LOGGER.warning(msg)
 
         conf_database = CONFIG.config['DATABASE']
         file = os.path.realpath(os.path.expanduser(conf_database['file']))